+++ test.lua
@@ -10,6 +10,8 @@ ROOT = fs.path(rootPath)
 LANG = 'zh-CN'
 TEST = true
 DEVELOP = true
+LOGPATH  = LOGPATH  or (ROOT .. '/log')
+METAPATH = METAPATH or (ROOT .. '/meta')
 
 collectgarbage 'generational'